project/backend/requirements.txt
fastapi==0.63.0
uvicorn==0.13.4
asyncpg==0.22.0
pytest==6.2.2
requests==2.25.1
SQLAlchemy==1.4.20
GeoAlchemy2==0.9.2
async-exit-stack==1.0.1
flake8==3.9.2
black==21.6b0
isort==5.9.1
async-generator
psycopg2
fastapi_pagination
alembic
shapely
fastapi_pagination
urllib3
geojson_pydantic